+++ file/src/pull.rs
@@ -12,8 +12,8 @@ use derive_more::Display;
 use xvc_core::{ContentDigest, XvcCachePath, XvcPath, XvcRoot};
 use xvc_ecs::{HStore, XvcStore};
 use xvc_logging::{watch, XvcOutputLine};
-use xvc_remote::{
-    remote::get_remote_from_store, StorageIdentifier, XvcStorage, XvcStorageGuid,
+use xvc_storage::{
+    storage::get_remote_from_store, StorageIdentifier, XvcStorage, XvcStorageGuid,
     XvcStorageOperations,
 };
 use xvc_walker::Glob;
