+++ storage/src/storage/gcs.rs
@@ -1,45 +1,40 @@
-use std::io::Write;
 use std::str::FromStr;
 use std::{env, fs};
 
-use crossbeam_channel::Sender;
-use futures::executor::block_on;
 use regex::Regex;
 use s3::creds::Credentials;
 use s3::{Bucket, Region};
 use serde::{Deserialize, Serialize};
-use tokio;
-use xvc_core::{XvcCachePath, XvcRoot};
+use xvc_core::XvcCachePath;
 use xvc_ecs::R1NStore;
 use xvc_logging::{watch, XvcOutputLine};
 
-use crate::remote::XvcStorageReceiveEvent;
+use crate::storage::XVC_REMOTE_GUID_FILENAME;
 use crate::{Error, Result, XvcStorage, XvcStorageEvent};
 use crate::{XvcStorageGuid, XvcStorageOperations};
 
 use super::{
     XvcStorageDeleteEvent, XvcStorageInitEvent, XvcStorageListEvent, XvcStoragePath,
-    XvcStorageSendEvent, XVC_REMOTE_GUID_FILENAME,
+    XvcStorageReceiveEvent, XvcStorageSendEvent,
 };
 
-pub fn cmd_new_minio(
+pub fn cmd_new_gcs(
     input: std::io::StdinLock,
-    output_snd: Sender<XvcOutputLine>,
-    xvc_root: &XvcRoot,
+    output_snd: crossbeam_channel::Sender<XvcOutputLine>,
+    xvc_root: &xvc_core::XvcRoot,
     name: String,
-    endpoint: String,
     bucket_name: String,
     region: String,
     remote_prefix: String,
 ) -> Result<()> {
-    let remote = XvcMinioRemote {
+    let remote = XvcGcsRemote {
         guid: XvcStorageGuid::new(),
         name,
         region,
         bucket_name,
         remote_prefix,
-        endpoint,
     };
+
     watch!(remote);
 
     let init_event = remote.init(output_snd.clone(), xvc_root)?;
@@ -50,7 +45,7 @@ pub fn cmd_new_minio(
         let event_e = xvc_root.new_entity();
         store.insert(
             store_e,
-            XvcStorage::Minio(remote.clone()),
+            XvcStorage::Gcs(remote.clone()),
             event_e,
             XvcStorageEvent::Init(init_event.clone()),
         );
@@ -61,16 +56,15 @@ pub fn cmd_new_minio(
 }
 
 #[derive(Clone, Debug, PartialOrd, Ord, PartialEq, Eq, Serialize, Deserialize)]
-pub struct XvcMinioRemote {
+pub struct XvcGcsRemote {
     pub guid: XvcStorageGuid,
     pub name: String,
     pub region: String,
     pub bucket_name: String,
     pub remote_prefix: String,
-    pub endpoint: String,
 }
 
-impl XvcMinioRemote {
+impl XvcGcsRemote {
     fn credentials(&self) -> Result<Credentials> {
         Credentials::new(
             Some(&env::var("XVC_REMOTE_ACCESS_KEY_ID").unwrap()),
@@ -83,14 +77,14 @@ impl XvcMinioRemote {
     }
 
     fn get_bucket(&self) -> Result<Bucket> {
-        // We'll just put guid file to endpoint/bucket/prefix/XVC_GUID_FILENAME
         let credentials = self.credentials()?;
         let region = Region::Custom {
-            region: self.region.clone(),
-            endpoint: self.endpoint.clone(),
+            region: self.region.to_owned(),
+            endpoint: "https://storage.googleapis.com".to_owned(),
         };
-        let bucket = Bucket::new(&self.bucket_name, region, credentials)?;
-        Ok(bucket.with_path_style())
+        let mut bucket = Bucket::new(&self.bucket_name, region, credentials)?;
+        bucket.set_listobjects_v1();
+        Ok(bucket)
     }
 
     async fn a_init(
@@ -302,7 +296,7 @@ impl XvcMinioRemote {
     }
 }
 
-impl XvcStorageOperations for XvcMinioRemote {
+impl XvcStorageOperations for XvcGcsRemote {
     fn init(
         &self,
         output: crossbeam_channel::Sender<xvc_logging::XvcOutputLine>,
